C++指针转换为uintptr_t类型时对应java的long类型

+++ app/src/main/cpp/jni/evo_player_jni.h
@@ -12,9 +12,9 @@ jstring get_ffmpeg_version(JNIEnv *env, jobject obj);
 jint play_video(JNIEnv *env, jobject obj,
                 jstring video_path, jobject surface);
 
-jint create_player(JNIEnv *env, jobject obj,
+jlong create_player(JNIEnv *env, jobject obj,
                    jstring video_path, jobject surface);
 
-void play(JNIEnv *env, jobject obj, jint player);
+void play(JNIEnv *env, jobject obj, jlong player);
 
 #endif //EVOPLAYER_EVO_PLAYER_JNI_H
